Solution to Check Alphabet Case

+++ Check Alphabet Case@@6788dd63490637cfc44e1b92/code.c
@@ -3,7 +3,7 @@
 int main() {
     char a;
     scanf("%c", &a);
-    if (((int)a) => a && ((int)a) <= z){
+    if (a => a && a <= z){
         printf("Uppercase");
     }
     else {
